# Optimize some functions in vp8.rs