This uses slices to index into the relevant portions of the array for top_pixels and edge_pixels, which helps decrease runtime by ~100%, as seen below:

```
test webp::vp8::bench::bench_edge_pixels                       ... bench:           8 ns/iter (+/- 0)
test webp::vp8::bench::bench_edge_pixels_new                   ... bench:           0 ns/iter (+/- 0)
test webp::vp8::bench::bench_top_pixels                        ... bench:           5 ns/iter (+/- 0)
test webp::vp8::bench::bench_top_pixels_new                    ... bench:           0 ns/iter (+/- 0)
```

I also tweaked avg2 to avoid u16 conversions. 
I added tests to cover all changes made, (including tests which fully cover the output of avg2 and avg3 to ensure that they remain consistent). The tests for edge_pixels and top_pixels also help demonstrate how the functions work.