@HeroicKatora 2020-02-14
Runtime of `0` means that llvm has constant inlined everything. You need to wrap both input and output for such small test in `test::black_box` to avoid this.