@philippeitis 2020-02-14
I added some black_box calls and got these benchmarks:

```
test webp::vp8::bench::bench_edge_pixels                       ... bench:           8 ns/iter (+/- 0)
test webp::vp8::bench::bench_edge_pixels_new                   ... bench:           1 ns/iter (+/- 0)
test webp::vp8::bench::bench_top_pixels                        ... bench:           5 ns/iter (+/- 0)
test webp::vp8::bench::bench_top_pixels_new                    ... bench:           1 ns/iter (+/- 0)
```