@philippeitis 2020-02-15
Old:
```
test webp::vp8::bench::bench_predict_4x4                       ... bench:         600ns/iter (+/- 87)
test webp::vp8::bench::bench_predict_4x4                       ... bench:         619 ns/iter (+/- 15)
test webp::vp8::bench::bench_predict_4x4                       ... bench:         585 ns/iter (+/- 44)
test webp::vp8::bench::bench_predict_4x4                       ... bench:         615 ns/iter (+/- 25)
test webp::vp8::bench::bench_predict_4x4                       ... bench:         639 ns/iter (+/- 63)
```

New:
```
test webp::vp8::bench::bench_predict_4x4                       ... bench:         589 ns/iter (+/- 26)
test webp::vp8::bench::bench_predict_4x4                       ... bench:         627 ns/iter (+/- 15)
test webp::vp8::bench::bench_predict_4x4                       ... bench:         626 ns/iter (+/- 7)
test webp::vp8::bench::bench_predict_4x4                       ... bench:         604 ns/iter (+/- 9)
test webp::vp8::bench::bench_predict_4x4                       ... bench:         583 ns/iter (+/- 22)
```

The benchmarks seem to favour the new version slightly, but given the high error, I can't make any conclusive statements on which version of avg2 is better, so I suppose it's a matter of personal taste.