@HeroicKatora 2020-02-15
The median absolute deviation of those samples (73 and 81 respectively) is several times the difference between their average estimates, I don't think we can draw any definitive conclusions based on those. Note that `avg2` is only used in two prediction methods and likely a small contribution to the total runtime. Also note that the input is not masked with `black_box` so maybe even some inopportune inlining of prediction methods and sample values is involved?