@fintelia 2020-02-15
Based on the [generated assembly](https://rust.godbolt.org/z/kLUG38) for the two versions, I'm rather skeptical that the new version is faster. If any differences are detectable beyond just being noise, then it looks like the new one will actually be a couple cycles _slower_.