@HeroicKatora 2020-02-15
The assembly of a single call isn't reallly important here. In a single function, of course the compiler will take advantage of the full register width and there is no advantage to keeping the values small. However, for the purpose of SIMDification it might require double the instruction count due to halving the number of vector elements. However, the bounds checks might present a problem for that optimization route (addressing this is the part of the change here).