@philippeitis 2020-02-15
Given the assembly and unreliability of the tests, I've decided to just revert the function.