@philippeitis 2020-02-15
Benchmarks for predict_bvepred consistently show a 6% speedup, with no margin of error. (Assembly is 30% smaller as well).
```
test webp::vp8::bench::bench_predict_bvepred                   ... bench:          16 ns/iter (+/- 0)
test webp::vp8::bench::bench_predict_bvepred_new               ... bench:          15 ns/iter (+/- 0)
```