@HeroicKatora 2020-02-15
> Benchmarks for predict_bvepred consistently show a 6% speedup, with no margin of error. (Assembly is 30% smaller as well).

Microbenchmarking like that can be very misleading. On that scale, you would need to have cycle counts, register pressure, and other CPU stats to mae any definitive statements as those are just a few hunder cycles. Please note that the use of `black_box` is also not quite correct. The compiler will search for constants or duplicate calculations. You need to ensure that *each* iteration gets 'fresh' inputs by wrapping them in `black_box` within the loop. The other ensures that the input is not completely constant folded but does not prevent the compiler from finding that each loop calculates the same result. It should read something like:

```rust
    fn bench_edge_pixels(b: &mut Bencher) {
        let v = make_sample_image();

        b.iter(|| {
            black_box(edge_pixels(black_box(&v), 5, 5, W * 2));
        });
    }
```