@philippeitis 2020-02-15
Updated edge_pixel benchmarks with more black boxing:
```
test webp::vp8::bench::bench_edge_pixels                       ... bench:           7 ns/iter (+/- 0)
test webp::vp8::bench::bench_edge_pixels_new                   ... bench:           2 ns/iter (+/- 0)
```