@HeroicKatora 2020-02-15
>  Updated edge_pixel benchmarks with more black boxing:

Indeed, the function alone does less bounds checks. This is not news. Your micro benchmarks are getting spammy though, these are of little value to the performance of the whole decoding pipeline. The point of the optimizer is to be able to prove some of these bounds checks unecessary and remove them by inlining those small utilities so those benchmarks add little knowledge for the performance of the prediction stage.