@philippeitis 2020-02-17
That's fair - I won't include any more micro-benchmarks, but the performance of predict_4x4 has improved by 15% - (the baseline time is 492ns +\- 5ns to 422ns +\- 4ns) with all of the changes I have included here - however, using only one optimized function at a time degrades performance (except for add_residue, which provides a 20ns improvement when used).