@HeroicKatora 2020-02-17
That's a good find! Providing more strict type information definitely seems to help here. Can you add a comment to the function (and/or the local copies) to explain this? Thanks for transferring the structural improvements to the other places as well.

And if you feel comfortable with it, I wouldn't mind a somewhat cleaned up commit history, force pushing allowed. Specifically, anything that makes it easier to cherry-pick the test suite and test the complete speedup of all changes together.