@philippeitis 2020-02-17
I've added comments and I've cleaned up the commit history. I've kept the optimizations in their separate commits, and the last commit contains all tests, optimizations and benchmarks you need to test the complete speed up.