@HeroicKatora 2020-02-19
Some numbers from my end as well:
```rust
// pre
test webp::vp8::test::bench_edge_pixels                        ... bench:           6 ns/iter (+/- 0)
test webp::vp8::test::bench_predict_4x4                        ... bench:         554 ns/iter (+/- 22)
test webp::vp8::test::bench_predict_bhepred                    ... bench:          12 ns/iter (+/- 0)
test webp::vp8::test::bench_predict_bldpred                    ... bench:          14 ns/iter (+/- 0)
test webp::vp8::test::bench_predict_brdpred                    ... bench:          16 ns/iter (+/- 2)
test webp::vp8::test::bench_predict_bvepred                    ... bench:          13 ns/iter (+/- 0)
test webp::vp8::test::bench_top_pixels                         ... bench:           4 ns/iter (+/- 0)

// post
test webp::vp8::test::bench_edge_pixels                        ... bench:           2 ns/iter (+/- 0)
test webp::vp8::test::bench_predict_4x4                        ... bench:         429 ns/iter (+/- 12)
test webp::vp8::test::bench_predict_bhepred                    ... bench:          10 ns/iter (+/- 0)
test webp::vp8::test::bench_predict_bldpred                    ... bench:          12 ns/iter (+/- 0)
test webp::vp8::test::bench_predict_brdpred                    ... bench:          16 ns/iter (+/- 1)
test webp::vp8::test::bench_predict_bvepred                    ... bench:           9 ns/iter (+/- 1)
test webp::vp8::test::bench_top_pixels                         ... bench:           1 ns/iter (+/- 0)
```